Fixed infinite recursion

+++ src/main/java/frc/team4069/robot/commands/OperatorDriveCommand.java
@@ -55,7 +55,7 @@ public class OperatorDriveCommand extends CommandBase {
                     * Math.PI;
 
             // Get the sign of the angle in order to calculate the direction to turn the wheels
-            double turnDirection = Math.signum((double) quickTurnDistanceMeters);
+            double turnDirection = Math.signum(quickTurnDistanceMeters);
             // Start turning at full speed in the direction of the sign
             driveBase.driveContinuousSpeed(turnDirection, 0);
         }
